Fix to clb function
Skip if time is lacking to timescale class

Fix file name for pro to be as for red

Add custom sx as option.  Do not save this to file.  Add flatten
command to test for nans.

Add fix_bad_files function to fix bad nc files (auto-generated
files do not necessarily have all the information)